# refactor: Native Model