--- a/packages/-ember-data/node-tests/docs/test-coverage.js
+++ b/packages/-ember-data/node-tests/docs/test-coverage.js
@@ -19,7 +19,7 @@ QUnit.module('Docs coverage', function(hooks) {
 
   QUnit.module('modules', function() {
     test('We have all expected modules', function(assert) {
-      assert.deepEqual(Object.keys(docs.modules), expected.modules, 'We have all modules');
+      assert.deepEqual(Object.keys(docs.modules).sort(), expected.modules, 'We have all modules');
     });
   });