--- a/packages/model/addon/-private/belongs-to.js
+++ b/packages/model/addon/-private/belongs-to.js
@@ -112,32 +112,20 @@ import { computedMacroWithOptionalParams } from './util';
   @return {Ember.computed} relationship
 */
 function belongsTo(modelName, options) {
-  let opts, userEnteredModelName;
-  if (typeof modelName === 'object') {
-    opts = modelName;
-    userEnteredModelName = undefined;
-  } else {
-    opts = options;
-    userEnteredModelName = modelName;
-  }
-
-  if (typeof userEnteredModelName === 'string') {
-    userEnteredModelName = normalizeModelName(userEnteredModelName);
-  }
-
   assert(
     'The first argument to belongsTo must be a string representing a model type key, not an instance of ' +
-      inspect(userEnteredModelName) +
+      inspect(modelName) +
       ". E.g., to define a relation to the Person model, use belongsTo('person')",
-    typeof userEnteredModelName === 'string' || typeof userEnteredModelName === 'undefined'
+    typeof modelName !== 'string'
   );
 
-  opts = opts || {};
+  modelName = normalizeModelName(modelName);
+  options = options || {};
 
   let meta = {
-    type: userEnteredModelName,
+    type: modelName,
     isRelationship: true,
-    options: opts,
+    options,
     kind: 'belongsTo',
     name: 'Belongs To',
     key: null,
@@ -151,7 +139,7 @@ function belongsTo(modelName, options) {
             `'${key}' is a reserved property name on instances of classes extending Model. Please choose a different property name for your belongsTo on ${this.constructor.toString()}`
           );
         }
-        if (Object.prototype.hasOwnProperty.call(opts, 'serialize')) {
+        if (Object.prototype.hasOwnProperty.call(options, 'serialize')) {
           warn(
             `You provided a serialize option on the "${key}" property in the "${this._internalModel.modelName}" class, this belongs in the serializer. See Serializer and it's implementations https://api.emberjs.com/ember-data/release/classes/Serializer`,
             false,
@@ -161,7 +149,7 @@ function belongsTo(modelName, options) {
           );
         }
 
-        if (Object.prototype.hasOwnProperty.call(opts, 'embedded')) {
+        if (Object.prototype.hasOwnProperty.call(options, 'embedded')) {
           warn(
             `You provided an embedded option on the "${key}" property in the "${this._internalModel.modelName}" class, this belongs in the serializer. See EmbeddedRecordsMixin https://api.emberjs.com/ember-data/release/classes/EmbeddedRecordsMixin`,
             false,