--- a/packages/model/addon/-private/model.js
+++ b/packages/model/addon/-private/model.js
@@ -7,6 +7,7 @@ import { DEBUG } from '@glimmer/env';
 import Ember from 'ember';
 
 import { RECORD_DATA_ERRORS, RECORD_DATA_STATE, REQUEST_SERVICE } from '@ember-data/canary-features';
+import { HAS_DEBUG_PACKAGE } from '@ember-data/private-build-infra';
 import {
   DEPRECATE_EVENTED_API_USAGE,
   DEPRECATE_MODEL_TOJSON,
@@ -1173,68 +1174,6 @@ class Model extends EmberObject {
     return this._internalModel.referenceFor('hasMany', name);
   }
 
-  /**
-   Provides info about the model for debugging purposes
-   by grouping the properties into more semantic groups.
-
-   Meant to be used by debugging tools such as the Chrome Ember Extension.
-
-   - Groups all attributes in "Attributes" group.
-   - Groups all belongsTo relationships in "Belongs To" group.
-   - Groups all hasMany relationships in "Has Many" group.
-   - Groups all flags in "Flags" group.
-   - Flags relationship CPs as expensive properties.
-
-   @method _debugInfo
-   @for Model
-   @private
-   */
-  _debugInfo() {
-    let attributes = ['id'];
-    let relationships = {};
-    let expensiveProperties = [];
-
-    this.eachAttribute((name, meta) => attributes.push(name));
-
-    let groups = [
-      {
-        name: 'Attributes',
-        properties: attributes,
-        expand: true,
-      },
-    ];
-
-    this.eachRelationship((name, relationship) => {
-      let properties = relationships[relationship.kind];
-
-      if (properties === undefined) {
-        properties = relationships[relationship.kind] = [];
-        groups.push({
-          name: relationship.kind,
-          properties,
-          expand: true,
-        });
-      }
-      properties.push(name);
-      expensiveProperties.push(name);
-    });
-
-    groups.push({
-      name: 'Flags',
-      properties: ['isLoaded', 'hasDirtyAttributes', 'isSaving', 'isDeleted', 'isError', 'isNew', 'isValid'],
-    });
-
-    return {
-      propertyInfo: {
-        // include all other mixins / properties (not just the grouped ones)
-        includeOtherProperties: true,
-        groups: groups,
-        // don't pre-calculate unless cached
-        expensiveProperties: expensiveProperties,
-      },
-    };
-  }
-
   notifyBelongsToChange(key) {
     this.notifyPropertyChange(key);
   }
@@ -2149,6 +2088,70 @@ const ID_DESCRIPTOR = {
 dependentKeyCompat(ID_DESCRIPTOR);
 
 Object.defineProperty(Model.prototype, 'id', ID_DESCRIPTOR);
+
+if (HAS_DEBUG_PACKAGE) {
+  /**
+   Provides info about the model for debugging purposes
+   by grouping the properties into more semantic groups.
+
+   Meant to be used by debugging tools such as the Chrome Ember Extension.
+
+   - Groups all attributes in "Attributes" group.
+   - Groups all belongsTo relationships in "Belongs To" group.
+   - Groups all hasMany relationships in "Has Many" group.
+   - Groups all flags in "Flags" group.
+   - Flags relationship CPs as expensive properties.
+
+   @method _debugInfo
+   @for Model
+   @private
+   */
+  Model.prototype._debugInfo = function() {
+    let attributes = ['id'];
+    let relationships = {};
+    let expensiveProperties = [];
+
+    this.eachAttribute((name, meta) => attributes.push(name));
+
+    let groups = [
+      {
+        name: 'Attributes',
+        properties: attributes,
+        expand: true,
+      },
+    ];
+
+    this.eachRelationship((name, relationship) => {
+      let properties = relationships[relationship.kind];
+
+      if (properties === undefined) {
+        properties = relationships[relationship.kind] = [];
+        groups.push({
+          name: relationship.kind,
+          properties,
+          expand: true,
+        });
+      }
+      properties.push(name);
+      expensiveProperties.push(name);
+    });
+
+    groups.push({
+      name: 'Flags',
+      properties: ['isLoaded', 'hasDirtyAttributes', 'isSaving', 'isDeleted', 'isError', 'isNew', 'isValid'],
+    });
+
+    return {
+      propertyInfo: {
+        // include all other mixins / properties (not just the grouped ones)
+        includeOtherProperties: true,
+        groups: groups,
+        // don't pre-calculate unless cached
+        expensiveProperties: expensiveProperties,
+      },
+    };
+  };
+}
 
 if (DEPRECATE_EVENTED_API_USAGE) {
   /**