--- a/packages/store/package.json
+++ b/packages/store/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@ember-data/canary-features": "3.28.0-alpha.0",
     "@ember-data/private-build-infra": "3.28.0-alpha.0",
+    "@glimmer/tracking": "^1.0.4",
     "@ember/string": "^1.0.0",
     "ember-cli-babel": "^7.26.3",
     "ember-cli-path-utils": "^1.0.0",