Now that our minimum Ember LTS is ~3.12 (and for current canary 3.16! although we didn't go that far here)~  🎉 **3.20** 🎉we can refactor Model to be a native class. (support for this began in 3.6 and decorator support began in 3.10)

Note: ~in 3.16 minimum support~ **we can embrace a @tracked world**, we should prepare :) (this also helps us prepare)

This PR does that refactor as minimally as possible. The few test changes were minor cleanups I did while investigating failures (not fixes, just made debugging easier) that seemed worthwhile to just keep.