@runspired 2020-06-13
cc @snewcomer if we can land this then we can probably do some nice things to make sure that folks calling `destroyRecord` today don't get surprised by `isDeleted` suddenly returning false when #7217 lands