@snewcomer 2020-09-06
https://github.com/ef4/ember-resource-metadata/pull/25
https://github.com/embermap/ember-data-storefront/pull/396