@runspired 2021-04-10
```
Benchmark Results Summary    

duration phase estimated regression +130ms [103ms to 149ms] OR +4.78% [3.78% to 5.47%]
Phase [navigationStart] => [start-find-all] phase no difference [-11ms to 17ms]
Phase [start-find-all] => [start-outer-materialization] phase no difference [-18ms to 13ms]
Phase [start-outer-materialization] => [stop-outer-materialization] phase estimated regression +132ms [118ms to 143ms] OR +18.93% [17.04% to 20.51%]
Phase [stop-outer-materialization] => [end-find-all] phase no difference [0ms to 0ms]
Phase [end-find-all] => [Test End] phase no difference [-1ms to 0ms]

```

Apparently this is a sizeable regression (~19%) over the .extends syntax. We should look to see if we've messed up anything obvious.